fix run-research black-hole crashes and remove per-row hot-path waste
Pipeline correctness (each one aborted or silently emptied run-research):
- register strategy_resource_usage.csv schema + anomaly alias; the resource
  ledger crashed every run right after cache_export with KeyError.
- treat NaN as missing in feature_matrix _is_missing: the canonical Parquet
  sidecar read path returns NaN for absent optional floats via pandas, which
  failed the finite-value contract (ATR_1d_asof_t and every optional float).
- import MarketDataContractError in features.matrix so the row-parse except
  clause no longer raises NameError that masks the real error.
- labels streaming reader now consumes the parquet-preferring typed state/
  future iterators instead of the header-only CSV stubs, which produced zero
  labels and aborted prediction with an empty-label join.
- atlas polars backend casts future ATR columns to Float64 so all-null future
  horizons no longer crash with `abs` on Null dtype (short/empty windows).

Performance (per-row hot paths, no behaviour change):
- artifacts writer and atlas row serialization use a shallow dataclass field
  read instead of dataclasses.asdict, dropping its recursive per-field deepcopy.
- state/future row parsers drop pandas.isna for cheap NaN-aware checks on the
  10M-row candle stream.

Net: full pipeline runs end-to-end; suite 15->7 pre-existing failures, all
remaining are stale test fixtures, no regressions.

+++ src/anomaly_science/atlas/polars_backend.py
@@ -390,9 +390,12 @@ def _build_polars_nature_atlas_rows(*, joined, config: AtlasConfig) -> tuple[Atl
 def _with_polars_horizon_columns(frame, *, horizon: int, config: AtlasConfig):
     pl = _import_polars_for_atlas_backend()
     suffix = f"{horizon}m"
-    future_return_atr = pl.col(f"future_return_atr_{suffix}")
-    future_max_atr = pl.col(f"future_max_atr_{suffix}")
-    future_min_atr = pl.col(f"future_min_atr_{suffix}")
+    # When a future horizon is entirely null (e.g. short windows with no resolved
+    # paths), Parquet/CSV yield a Null-dtype column on which `.abs()` is invalid.
+    # Cast to Float64 so null stays null but numeric ops (abs/compare) are defined.
+    future_return_atr = pl.col(f"future_return_atr_{suffix}").cast(pl.Float64)
+    future_max_atr = pl.col(f"future_max_atr_{suffix}").cast(pl.Float64)
+    future_min_atr = pl.col(f"future_min_atr_{suffix}").cast(pl.Float64)
     stop_first = pl.col(f"barrier_resolution_{suffix}") == BARRIER_RESOLUTION_STOP_LOSS_FIRST
     missing = future_return_atr.is_null() | future_max_atr.is_null() | future_min_atr.is_null()
     upside = future_max_atr >= config.outcome_continuation_threshold_atr
